Minor fixes
AbilityBattleEffects fix arg
Gimme the two fucking bikes mate
Sandstream prevents damage from sandstorm
Wicked blow gets cross chop animation
Ferroseed Leech Seed
Kabutops Stone Axe
Yawn affected by protect
update readme
Cleanup
added comment cmd_jumpifaffectedbyprotect
remove match call crap
Fix space center battle lock out
Mirage island always present

+++ src/battle_util.c
@@ -4129,7 +4129,7 @@ static u8 ForewarnChooseMove(u32 battler)
     free(data);
 }
 
-u8 AbilityBattleEffects(u8 caseID, u8 battler, u16 ability, u8 special, u16 moveArg)
+u8 AbilityBattleEffects(u8 caseID, u8 battler, u16 ability, u16 special, u16 moveArg)
 {
     u8 effect = 0;
     u32 speciesAtk, speciesDef;
